backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
nltk==3.9.1
tensorflow==2.18.0
Pillow==11.1.0
numpy==1.26.4
